Write a program that multiplies two positive numbers,Usage: mul num1 num2,num1 and num2 will be passed in base 10,Print the result, followed by a new line,If the number of arguments is incorrect, print Error, followed by a new line, and exit with a status of 98,num1 and num2 should only be composed of digits. If not, print Error, followed by a new line, and exit,with a status of 98,You are allowed to use more than 5 functions in your file

+++ 0x0C-more_malloc_free/101-mul.c
@@ -49,12 +49,12 @@ int _atoi(const char *s)
  *
  * Return: 0
  */
-void print_int(unsigned long int n)
+void print_int(int n)
 {
-	unsigned long int divisor = 1, i, response;
+	int divisor = 1, i, response;
 
 	for (i = 0; n / divisor > 9; i++, divisor *= 10)
-	;
+		;
 
 	for (; divisor >= 1; n %= divisor, divisor /= 10)
 	{
